Don't die if the sim robot doesn't exit

+++ run_sim.rb
@@ -3,10 +3,13 @@ require 'fileutils'
 require 'mkmf'
 
 has_code = File.exists?(ARGV[0])
+has_sim_code = File.exists?(ARGV[1])
 abort "Error: Robot code doesn't exist in project #{ARGV[0]}" unless has_code
+abort "Error: Sim Robot code doesn't exist in project #{ARGV[0]}" unless has_sim_code
 abort "Error: Ant not found. Try installing with 'brew install ant'." unless find_executable 'ant'
 
 cur_dir = Dir.pwd
+puts cur_dir
 Dir.chdir(ARGV[0])
 system('ant jar')
 
@@ -30,11 +33,18 @@ Dir.chdir('classes')
 `jar -xf ../FRCUserProgram.jar`
 FileUtils.cp "META-INF/MANIFEST.MF", "META-INF/MANIFEST.MF.old"
 
-`jar -xf ../FakeWPILib.jar`
+#{}`jar -xf ../FakeWPILib.jar`
+`cp -r ../../bin/ .`
 FileUtils.cp "META-INF/MANIFEST.MF.old", "META-INF/MANIFEST.MF"
 FileUtils.rm "META-INF/MANIFEST.MF.old"
 
+# Copy sim robot class files
+puts `pwd`
+`cp -r #{cur_dir}/#{ARGV[1]}/bin/ .`
+
 Dir.chdir('..')
+puts `find . | grep Robot`
+
 `jar -cmvf classes/META-INF/MANIFEST.MF to_sim.jar -C classes .`
 
 system("java -jar to_sim.jar")
